requirements.txt
fastapi
uvicorn[standard]
neo4j
python-dotenv
pydantic
neo4j-cypher